Ensure `parseInt` is called with the correct radix.
Commander passes both the user specified value and the default value
into the callback for `.option`. This means that we were calling
`parseInt(<user value>, <cpu count - 1>)`.

When the system cpu count was higher than 4 (the number of cpus that the
`concurrency` test specifies) the test would pass (which is why CI was
passing), however on my MacBook Pro with only 4 cpus we run
`parseInt('4', 3)` which results in `NaN` and therefore fails the
`concurrency` test.