Fix: Improve handling of removed hash pairs
- Ensure that linesFrom always returns one line, so one-line inline
changes don't cause lines to disappear.
- Add a couple of sanity checks around whitespace errors that may be
introduced by splicing in content. Specifically, if the line is empty
and we made a change, remove it; and if there is whitespace surrounding
our edit, trim the preceeding whitespace away.
- Turns on Glimmer preprocess option to preserve more whitespace in
printed ASTs. (Currently passing in the prop is a no-op; depends on
Glimmer PR that's in review.)