Only push last modification for specific replacement range.
When replacing one node with an array all of the replacement nodes have
the same location information. Prior to these changes, we would attempt
to do all of the operations which results in incorrect application of
the modifications. This change ensures that we only do the _last_ one,
which will include the other changes.

This fixes a somewhat long standing issue (the `QUnit.todo` changed to a
normal passing `QUnit.test`) where you could not push a hash pair into a
previously empty hash.