Ensure proxy mapping is maintained during set/delete.
Previously, when a mutation of an existing property occurred the proxy
system that we had in place would never reflect those changes
externally. This meant that some operations (initially identified as
array operations) would be lost when moving existing properties
unchanged.

This ensures that the underlying proxy / property detection system will
reflect the changes that are made when accessing the properties
subsequently.