Ensure chained inverses are properly printed.
Previously, when a chained inverse was updated we would only reprint the
inverse's `program` node. This meant that if you used a chained inverse
**and** a terminal inverse that reprinting the entire statement would
loose parts of the original template.

The reason we previously only printed the inverse's `program` was
because in the case where the inverse itself is chained, there is
actually no valid "end source" for that nested `BlockStatement` (😩).

The fix here is to "mark" the `BlockStatement` that is part of a chained
inverse _before_ attempting to print it, and when printing chained
inverse `BlockStatement`'s recursively avoid attempting to emit an
`endSource` (because there is none).