Ensure AttrNode and HashPair values are preserved if changing name/key.
Prior to this change, we retrieved the original source for a
`HashPair`/`AttrNode` value via a regular expression. The regexp is
needed in order to capture the actual source of the `key` / `name`
because there is no location information in the AST, however using a
regexp for matching the value was causing needless issues. The
`AttrNode`/`HashPair` `.value` will _always_ have its own location
information, and we should use that instead of a regexp to build up the
original source of the value.