Avoid issues with non-posix paths with globby
globby uses the fast-glob package, which states this in its README (RE:
patterns):

> ⚠️ Always use forward-slashes in glob expressions (patterns and ignore
> option). Use backslashes for escaping characters.

Previously, we were using `path.join` which will use the OS specific
separator and therefore fail to match files when invoked with a
directory path directly.

globby has an explicit API that allows us to avoid doing this
`path.join` at all (`expandDirectories`). This changes to use that,
remove the manual joining.

---

This also corrects for a bug in globby (to be reported/fixed separately)
where the `cwd` option is defaulted to `process.cwd()` _without_ posix
correction, and then subsequently an assertion is made that a given path
"starts with" the cwd.