Refactor test assertion to be platform agnostic.
globby _always_ returns posix style paths (regardless of platform).
In order to search for the "file with the error" in the output, we need
to convert the `path.join` result into posix style paths. We (and
globby) use `slash` for this in other areas, so it is used here for
simplicity and to match the logic globby would use.