Correct valueless AttrNode location munging.
simple-html-tokenizer interprets `<div data-foo></div>` as `<div
data-foo=""></div>`. In other words, valueless attributes are treated as
attributes with a `TextNode` value of `""`. Unfortunately, due to the
issue identified in glimmerjs/glimmer-vm#953 the AST that is returned
from `@glimmer/syntax` has the wrong end location for that `TextNode`
(it thinks the node is 1 char, but in reality it is 0 chars since it
doesn't actually exist) so we have to fix the AST location information
before attempting to run our transformations.

The original work around corrected the location information of *any*
`AttrNode` that had a `TextNode` with a `""` value. In retrospect this
is clearly incorrect, as it is perfectly valid to have authored your
template as `<div data-foo=""></div>`. The fix introduced in this commit
ensures that we only correct the location information if the `AttrNode`
itself was **actually** valueless (e.g. had no equal sign).