Ensure attributes, params, hash are only joined with whitespace.
A bug was fixed recently that was caused by improperly detecting
the join value for attributes. That issue was corrected (along with
tests to ensure no regressions), but upon reflection it seemed
completely bizarre that we would _ever_ allow the autodetected join
value to be something other than whitespace.

This commit inserts a fallback to ensure that if we ever do happen to
auto-detect a non-whitespace value as the join value, that we instead
use `' '`. There are no known scenarios where we can actually **hit**
this code, but it still seems reasonable to add in order to make this a
bit safter by default.