Make AttrNode responsible for custom quoting requirements.
The location information (as of @glimmer/syntax@0.47.9) for a `TextNode`
that is the sole value of an `AttrNode` _includes_ any surrounding
quotes. This should be fixed upstream, such that calculating the source
for the `TextNode` accounts for the actual text content (not the
surrounding formatting).

The primary changes here are ensuring that the only thing "aware" of
quotedness (or lack thereof) of an `AttrNode` value _is_ the `AttrNode`.
This is accomplished by adding an additional check to the AST
fixup prepass that we already do. This new fixup updates the location
information when a `TextNode` is an `AttrNode`'s `.value`.

In addition, this commit takes advantage of new capabilities in
`@glimmer/syntax` (passing the `path` in addition to the `node` while
visiting the AST nodes).