Expose metadata (isValueless / quoteType) on AttrNode's
The fixes introduced in a589561e caused issues for
ember-template-lint@2.x branch (which uses ember-template-recast as its
parser). The underlying reason for the issue is that ember-template-lint
was _basically_ relying on the bug fixed (e.g. it was looking at
`AttrNode`'s `.value` and doing `sourceForLoc` on that node, then
checking if it starts with a quote) .

In order for those ember-template-lint rules to function (`quotes` and
`no-quoteless-attribute`) they need to know if a given attribute had a
value (aka was it `<div data-test-blah></div>` vs `<div
data-test-blah="thing"></div>`) and what type of quotes were used
(single, double, none, etc).

This adds to the existing `AttrNode` interface, such that it is:

```ts
export interface AttrNode extends BaseNode {
  type: 'AttrNode';
  name: string;
  value: TextNode | MustacheStatement | ConcatStatement;
  isValueless: boolean;
  quoteType: '"' | "'" | null;
}
```