Fix issue when swapping ElementNode from self closing with attributes to block.
The original fix (that landed in 4.1.1) assumed that it was _always_
acceptable to remove the `postTagWhitespace` when moving from self
closing to having children. Unfortunately, that only worked because the
specific test cases added didn't use attributes (or block params).

This fix builds on a test provided by
[@lukemelia](https://github.com/lukemelia) and adds a few more tests
(for various combos of other argument types) and ensures that the
resulting whitespace is "intuitive" (yes, I know this is subjective, but
I made the changes so I get to say it is intuitive :P ).