Failing test demonstrating extraneous whitespace after hash.
This demonstrates a bug in glimmer-vm's AST override handling.
When we are writing a new AST node (e.g. a `MustacheStatement`) that
contains an AST node that we already have raw source for (e.g. an
existing `SubExpression`'s `Hash`) the glimmer-vm printer always ensures
that whitespace is included when it is needed (e.g. when the buffer for
a `MustacheStatement` is `{{foo` you **must** add a whitespace before
you can begin printing hash value). Unfortunately, when the node that we
already have source for has a value of `''` (empty string) the guard in
glimmer-vm's printer that ensures whitespace is added is wrong. In that
case it is perfectly fine to **not** add that whitespace.

Identified by ember-angle-brackets-codemod.