Change isSynthetic implementation
Starting from @glimmer/syntax@0.66.0, `node.loc.source` is always
`(synthetic)` when a node is created with a `builder`.

A synthetic node is a node created without any location.

So this commit changes isSynthetic implementation to rely on
`node.loc.{start,end}` instead of `node.loc.source`. This is a
workaround to the change in @glimmer/syntax.