Make printing more robust to manually created glimmer nodes
Fixes https://github.com/ember-template-lint/ember-template-recast/issues/586

Previously, calling print on node that came from our parse(), but was
not the top-level node, would cause infinite recursion. (It would have
NodeInfo, but not a registered parse result).

This fixes the infinite loop, by just storing a reference to the
top-level parse result in the node info, and using that.

Importantly, this successfully preserves formatting for any sub-nodes we
did parse ourselves, or any user-created nodes that require custom
printing.