Add `lib` field in tsconfig.json
As a result, we restrict build-in type declarations to es2019 and dom.

Adding `es2019` and `dom` groups to the list of built-in type
declarations is the minimum set we need.

Why do we need this change?

In a follow-up, we will add `es2020.string` to the list. It will allow
us to use String.matchAll which is also supported in Nodejs 12.